Darwin: Dead code removal, Code cleanup, Added launcher Imported changes from xorg-server-1.2-apple to make master more current wrt file layout, build system changes, and dead code removal.

+++ hw/darwin/quartz/xpr/appledri.c
@@ -174,7 +174,7 @@ ProcAppleDRIAuthConnection(
     rep.authenticated = 1;
 
     if (!DRIAuthConnection( screenInfo.screens[stuff->screen], stuff->magic)) {
-        ErrorF("Failed to authenticate %u\n", stuff->magic);
+        ErrorF("Failed to authenticate %u\n", (unsigned int)stuff->magic);
         rep.authenticated = 0;
     }
     WriteToClient(client, sizeof(xAppleDRIAuthConnectionReply), (char *)&rep);
